Refining SearchItemServlet.java, main.jsp, item_detail.jsp, and search_items.jsp
